Add coefficients to symmetry functions
Added coefficients to symmetry operation symbols in print_header() and print_table() functions as a convenience for the user. This makes it easier for the user to know the order, type, and number of symmetry operations for a point group without having to pull out character tables from another source.